+++ compiler/src/jdk.graal.compiler/src/jdk/graal/compiler/lir/aarch64/AArch64GHASHProcessBlocksOp.java
@@ -65,7 +65,7 @@ import jdk.vm.ci.meta.AllocatableValue;
 import jdk.vm.ci.meta.Value;
 
 // @formatter:off
-@SyncPort(from = "https://github.com/openjdk/jdk/blob/ce8399fd6071766114f5f201b6e44a7abdba9f5a/src/hotspot/cpu/aarch64/stubGenerator_aarch64.cpp#L6168-L6302",
+@SyncPort(from = "https://github.com/openjdk/jdk/blob/be2b92bd8b43841cc2b9c22ed4fde29be30d47bb/src/hotspot/cpu/aarch64/stubGenerator_aarch64.cpp#L6181-L6315",
           sha1 = "84b96e679b2ff5dc836da5c28fbbc779b5320a2b")
 @SyncPort(from = "https://github.com/openjdk/jdk/blob/12358e6c94bc96e618efc3ec5299a2cfe1b4669d/src/hotspot/cpu/aarch64/macroAssembler_aarch64_aes.cpp#L285-L691",
           sha1 = "1cd41d8f202ebe127aa31053ab3c6851f3900034")