+++ espresso/src/com.oracle.truffle.espresso/src/com/oracle/truffle/espresso/nodes/quick/invoke/InvokeStaticQuickNode.java
@@ -56,4 +56,8 @@ public final class InvokeStaticQuickNode extends InvokeQuickNode {
         Object[] args = getArguments(frame);
         return pushResult(frame, invokeStatic.execute(args));
     }
+
+    public void initializeResolvedKlass() {
+        invokeStatic.getStaticMethod().getDeclaringKlass().safeInitialize();
+    }
 }